Refactor: #36 - 회원 엔티티 리팩토링
- Gender, Role Enum 클래스를 Member 엔티티 클래스 내부로 옮겼습니다.
