--- src/main/java/com/fourfourfour/damoa/api/member/enums/Gender.java
@@ -1,7 +1,0 @@
-package com.fourfourfour.damoa.api.member.enums;
-
-public enum Gender {
-
-    MALE, FEMALE, ETC;
-
-}
